extract tensor helper microcrate and wire CLI
