+++ crates/bitnet-tensor-utils/Cargo.toml
@@ -1,0 +1,20 @@
+[package]
+name = "bitnet-tensor-utils"
+version.workspace = true
+edition.workspace = true
+license.workspace = true
+repository.workspace = true
+homepage.workspace = true
+authors.workspace = true
+keywords.workspace = true
+categories.workspace = true
+description = "SRP microcrate for extracting and inspecting BitNet tensors"
+rust-version.workspace = true
+
+[dependencies]
+anyhow.workspace = true
+bitnet-common = { path = "../bitnet-common", version = "0.2.1-dev" }
+candle-core.workspace = true
+
+[lints]
+workspace = true
